+++ src/dcp_tools/custom_data/__init__.py
@@ -1,0 +1,5 @@
+from dcp_tools.custom_data.data_management import CustomDataManager
+
+__all__ = [
+    "CustomDataManager",
+]